image-resizer/requirements.txt
boto3
pillow==8.0.0
mysql-connector-python==8.0.11